pyproject.toml
[tool.poetry]
name = "theridian"
version = "0.1.0"
description = "Theridian DRF API with ETL system"
authors = ["OS <@github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
# 3.13 still has many breaking changes
python = ">=3.10,<3.13"
django = "^5.0.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
celery = "^5.3.4"
redis = "^5.0.1"
dagster = "^1.11.0"
dagster-webserver = "^1.11.0"
dagster-postgres = "^0.27.0"
dagster-celery = "^0.27.0"
#  pika = "^1.3.2"
#  gunicorn = "^21.2.0"
#  whitenoise = "^6.6.0"
django-prometheus = "^2.3.1"
prometheus-client = "^0.19.0"
structlog = "^23.2.0"
django-structlog = "^8.0.0"
sentry-sdk = "^1.38.0"
dj-database-url = "^2.1.0"
django-filter = "^23.5"
drf-spectacular = "^0.27.0"
django-health-check = "^3.17.0"
django-redis = "^6.0.0"
# Production / Linux Only
bjoern = { version = "^3.2.0", optional = true }
colored = "^2.3.0"
pandas = "^2.3.0"
django-celery-beat = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
flake8-django = "^1.4"
mypy = "^1.7.1"
django-stubs = "^4.2.7"
pre-commit = "^3.6.0"
coverage = "^7.3.2"
django-debug-toolbar = "^4.2.0"

[tool.poetry.extras]
bjoern = ["bjoern"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = "core,apps"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]

[tool.coverage.run]
source = "."
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "*/settings/*",
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings.production"